Added token.getWithoutPrompt
Resolves: OKTA-94707
