+++ lib/clientBuilder.js
@@ -66,7 +66,7 @@ function OktaAuthBuilder(args) {
   };
 
   sdk.idToken = {
-    authorize: util.bind(token.getIdToken, sdk, sdk),
+    authorize: util.bind(token.getIdToken, sdk, sdk), // deprecated for sessionToken
     verify: util.bind(token.verifyIdToken, sdk, sdk),
     refresh: util.bind(token.refreshIdToken, sdk, sdk),
     decode: util.bind(token.decodeToken, sdk) // deprecated
@@ -78,6 +78,7 @@ function OktaAuthBuilder(args) {
   };
 
   sdk.token = {
+    getWithoutPrompt: util.bind(token.getWithoutPrompt, sdk, sdk),
     decode: util.bind(token.decodeToken, sdk)
   };
 }
