+++ lib/token.js
@@ -482,8 +482,19 @@ function getIdToken(sdk, oauthOptions, options) {
   }
 }
 
+function getWithoutPrompt(sdk, oauthOptions, options) {
+  var oauthParams = util.clone(oauthOptions) || {};
+  util.extend(oauthParams, {
+    prompt: 'none',
+    responseMode: 'okta_post_message',
+    display: null
+  });
+  return getIdToken(sdk, oauthParams, options);
+}
+
 module.exports = {
   getIdToken: getIdToken,
+  getWithoutPrompt: getWithoutPrompt,
   refreshIdToken: refreshIdToken,
   decodeToken: decodeToken,
   verifyIdToken: verifyIdToken
